bigint: Split `Modulus` (and `PartialModulus`) into a submodule.
Better encapsulate `Modulus` and `PartialModulus`.

`git diff HEAD^1:src/arithmetic/bigint.rs src/arithmetic/bigint/modulus.rs`
